adding loading to search bar
